backend/requirements.txt
fastapi
uvicorn
faster-whisper
transformers
sentencepiece
ffmpeg-python
python-dotenv
srt
setuptools<81